backend/crudapi/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1
django-ninja==1.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
